+++ src/lib/Scenes/City/CityPicker.tsx
@@ -2,7 +2,7 @@ import { Box, color, Flex, Sans, Separator, Serif, space } from "@artsy/palette"
 import { dimensions, screen } from "lib/data/ScreenSizes/screenSizes"
 import { CircleWhiteCheckIcon } from "lib/Icons/CircleWhiteCheckIcon"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
-import { Schema, screenTrack } from "lib/utils/track"
+import { Schema, screenTrack, track } from "lib/utils/track"
 import React, { Component } from "react"
 import { Dimensions, NativeModules, TouchableOpacity } from "react-native"
 import styled from "styled-components/native"
@@ -22,8 +22,6 @@ const cityList = cities.map(city => city.name)
 @screenTrack(() => ({
   context_screen: Schema.PageNames.CityPicker,
   context_screen_owner_type: Schema.OwnerEntityTypes.CityGuide,
-  context_screen_owner_slug: "",
-  context_screen_owner_id: "",
 }))
 export class CityPicker extends Component<Props, State> {
   constructor(props: Props) {
@@ -65,6 +63,12 @@ export class CityPicker extends Component<Props, State> {
     )
   }
 
+  @track(() => {
+    return {
+      action_name: Schema.ActionNames.BMWLogo,
+      action_type: Schema.ActionTypes.Tap,
+    } as any
+  })
   navigateToBMWArtGuide() {
     const { sponsoredContentUrl } = this.props